:art: Refactor Checkbox
