+++ src/stories/Form/Checkbox/index.spec.tsx
@@ -6,7 +6,7 @@ describe("#Checkbox", () => {
   it("should call onChange event when clicked", () => {
     const onChangeMock = jest.fn();
     const { getByLabelText } = render(
-      <Checkbox id="test-checkbox" checked={false} onChange={onChangeMock}>
+      <Checkbox id="test" name="test" checked={false} onChange={onChangeMock}>
         Test Checkbox
       </Checkbox>
     );
@@ -15,12 +15,11 @@ describe("#Checkbox", () => {
     fireEvent.click(checkbox);
 
     expect(onChangeMock).toHaveBeenCalledTimes(1);
-    expect(onChangeMock).toHaveBeenCalledWith(true);
   });
 
   it("should render with icon when withIcon is true", () => {
     const { container } = render(
-      <Checkbox id="test-checkbox" checked={false} withIcon>
+      <Checkbox id="test" name="test" checked={false} withIcon>
         Test Checkbox
       </Checkbox>
     );
@@ -30,7 +29,7 @@ describe("#Checkbox", () => {
 
   it("should render with focus styles when withFocus is true", () => {
     const { container } = render(
-      <Checkbox id="test-checkbox" checked={false} withFocus>
+      <Checkbox id="test" name="test" checked={false} withFocus>
         Test Checkbox
       </Checkbox>
     );
@@ -44,7 +43,7 @@ describe("#Checkbox", () => {
 
   it("should render text children as Text component", () => {
     const { container, getByText } = render(
-      <Checkbox id="test-checkbox" checked={false}>
+      <Checkbox id="test" name="test" checked={false}>
         Text Children
       </Checkbox>
     );
@@ -56,7 +55,7 @@ describe("#Checkbox", () => {
   it("should render non-text children as React elements", () => {
     const iconElement = <div data-testid="icon">Icon Element</div>;
     const { getByTestId } = render(
-      <Checkbox id="test-checkbox" checked={false}>
+      <Checkbox id="test" name="test" checked={false}>
         {iconElement}
       </Checkbox>
     );
