+++ src/stories/Form/Checkbox/type.ts
@@ -1,0 +1,10 @@
+import { InputHTMLAttributes, ReactNode } from "react";
+
+export interface CheckboxProps extends InputHTMLAttributes<HTMLInputElement> {
+  id: string;
+  name: string;
+  checked: boolean;
+  children: string | ReactNode;
+  withIcon?: boolean;
+  withFocus?: boolean;
+}
